third_party/WebKit/LayoutTests/fast/shapedetection/shapedetection-empty-input.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script>

// This test verifies FaceDetector.detect() with an empty HTMLImageElement.
var createTestForEmptyInput = function(detectorName) {
  async_test(function(t) {
    var image = new Image();
    var detector = eval("new " + detectorName + "();");
    var tryDetection = function() {
      detector.detect(image)
          .then(detectionResult => {
            assert_equals(detectionResult.length, 0);
            t.done();
          })
          .catch(error => {
            assert_unreached("detect() rejected with error: " + error)
          });
    };
    image.onload = tryDetection;
    image.onerror = tryDetection;
    image.src = "";
  }, "Detectors should resolve empty images with empty results.");
};

generate_tests(createTestForEmptyInput, [
  [ "Face", "FaceDetector" ], [ "Barcode", "BarcodeDetector" ]
]);

</script>
